# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareerPath AI Architect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analysis:
    post:
      operationId: runAnalysis
      tags: [analysis]
      summary: Run AI career analysis
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Optional key for safely retrying analysis creation without creating duplicate analyses. Reuse the same value for retries of the same request.
          schema:
            type: string
            minLength: 1
            maxLength: 128
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerAnalysis"
        "400":
          description: Incomplete profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analysis/latest:
    get:
      operationId: getLatestAnalysis
      tags: [analysis]
      summary: Get latest analysis
      responses:
        "200":
          description: Latest analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerAnalysis"
        "404":
          description: No analysis yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CareerAnalysis:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        targetRole:
          type: string
        readinessScore:
          type: integer
          minimum: 0
          maximum: 100
        readinessSubScores:
          $ref: "#/components/schemas/ReadinessSubScores"
        profileSummary:
          type: string
        currentStrengths:
          type: string
        currentStrengthsStructured:
          type: array
          items:
            $ref: "#/components/schemas/StructuredInsight"
        skillGaps:
          type: string
        skillGapsStructured:
          type: array
          items:
            $ref: "#/components/schemas/StructuredInsight"
        experienceGaps:
          type: string
        qualificationGaps:
          type: string
        certificationRecommendations:
          type: string
        suggestedProjects:
          type: string
        jobProgressionLadder:
          type: string
        immediateActions:
          type: string
        immediateActionsStructured:
          type: array
          items:
            $ref: "#/components/schemas/StructuredInsight"
        year1Priorities:
          type: string
        year2To3Plan:
          type: string
        year4To5Plan:
          type: string
        roadmapPhases:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapPhase"
        learningRecommendations:
          type: array
          items:
            $ref: "#/components/schemas/LearningRecommendationGroup"
        modelName:
          type: string
        promptVersion:
          type: string
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        latencyMs:
          type: integer
        profileSnapshot:
          $ref: "#/components/schemas/AnalysisProfileSnapshot"
        createdAt:
          type: string
      required: [id, userId, targetRole, readinessScore, readinessSubScores, profileSummary, currentStrengths, currentStrengthsStructured, skillGaps, skillGapsStructured, experienceGaps, qualificationGaps, certificationRecommendations, suggestedProjects, jobProgressionLadder, immediateActions, immediateActionsStructured, year1Priorities, year2To3Plan, year4To5Plan, roadmapPhases, learningRecommendations, modelName, promptVersion, inputTokens, outputTokens, latencyMs, profileSnapshot, createdAt]
    ReadinessSubScores:
      type: object
      properties:
        profile:
          type: integer
        skills:
          type: integer
        experience:
          type: integer
        education:
          type: integer
        certifications:
          type: integer
      required: [profile, skills, experience, education, certifications]
    StructuredInsight:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        priority:
          type: string
          enum: [High, Medium, Low]
        category:
          type: string
        skillCode:
          type: string
        skillLabel:
          type: string
      required: [title, detail]
    LearningRecommendationGroup:
      type: object
      properties:
        sourceType:
          type: string
          enum: [skill-gap, roadmap-phase]
        sourceId:
          type: string
        sourceLabel:
          type: string
        priority:
          type: string
          enum: [High, Medium, Low]
        timeframe:
          type: string
        skillCode:
          type: string
        skillLabel:
          type: string
        courses:
          type: array
          items:
            $ref: "#/components/schemas/LearningCourseRecommendation"
      required: [sourceType, sourceId, sourceLabel, courses]
    LearningCourseRecommendation:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        title:
          type: string
        url:
          type: string
          format: uri
        skills:
          type: array
          items:
            type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        cost:
          type: string
          enum: [free, freemium, paid]
        durationHours:
          type: integer
        format:
          type: string
          enum: [course, specialization, certificate, tutorial, playlist]
        description:
          type: string
        matchedSkills:
          type: array
          items:
            type: string
        matchReason:
          type: string
        score:
          type: integer
      required: [id, provider, title, url, skills, level, cost, durationHours, format, description, matchedSkills, matchReason, score]
    AnalysisProfileSnapshot:
      type: object
      properties:
        currentRole:
          type: ["string", "null"]
        totalExperienceMonths:
          type: ["integer", "null"]
        industry:
          type: ["string", "null"]
        careerLevel:
          type: ["string", "null"]
        weeklyLearningMinutes:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        workExperienceCount:
          type: integer
        educationCount:
          type: integer
        certificationCount:
          type: integer
      required: [currentRole, totalExperienceMonths, industry, careerLevel, weeklyLearningMinutes, skills, workExperienceCount, educationCount, certificationCount]
...
    RoadmapPhase:
      type: object
      properties:
        label:
          type: string
        timeframe:
          type: string
        actions:
          type: array
          items:
            type: string
        focus:
          type: string
      required: [label, timeframe, actions, focus]
...